Improvements in studio downloading
When StudioFlags aren't ON, we are downloading the entire string of
PRODUCTORAS into the same "<studio>" tag. Meaning that multiple studios
are seen by just one with a long name in XBMC. This behaviour has been
improved, making each studio be in it's own "<studio>" tag.

When StudioFlags are ON, we are downloading just first one in the list,
so that problem doesn't arise here. But there is other one a bit annoying.

Sometimes there is a string like "Coproducción EEUU-Alemania-Canadá;" [1]
preceding the actual first studio. We were downloading also that info,
and placing it into the "<studio>" tag, so thumbs from StudioFlags don't
work as expected. I've fixed this making the regexp check if that first
string exists (always ends with "; ") and removing it if that's true,
if it isn't, it will work as before.

[1] Example from: http://www.filmaffinity.com/es/film543392.html