Replaced credits regexp
Credits (GUIONISTAS in Filmaffinity), before this commit were extracted
as a full string, as it was only one screenwriter even when there are
more. So I replaced those regexps to split them in different <credit>
tags, same way as we do with directors.

There is, however, a small problem here, there are some films in
Filmaffinity which have as screenwriter the original book or story that
inspired the film. Examples are like "(Novela: Ben Mezrich)" [1]. Right
now, I am getting them also as another "<credit>" tag (and removing
parenthesis). But I don't know if this is the correct thing to do, or
if we could better left them out of credits list, so this is not
definitive.

[1]: http://www.filmaffinity.com/es/film577699.html