Added AlpacinePosters support
Issue#2: https://github.com/MaDDoGo/metadata.filmaffinity.com/issues/2

Added dowloading of Alpacine thumbs. It has been implemented with a
GoogleToAlpacine function that works the same way as  GoogleToIMDB. This
probably can be better done using the internal alpacine.com searching,
but works well this way, and it's just a first version.

After obtaining Alpacine's ID, we use another function to download all
available thumbs. This is pretty simple: inside "carteles/" page in
alpacine.com, there are small images of the same HighRes thumbs, the
name of this smaller ones is the same as the big ones with just an
"-100" string at the end [1]. So if we remove that we are going to got
the big poster.

There was already a conditional inside settings.xml called
EnableAlpacinePosters that wasn't being used, so I took that one to
enable or disable these downloads. It is checked before gettint into
GoogleToAlpacine, so we do not even go googling if it is disabled.

[1] Example:
In http://alpacine.com/pelicula/22694/carteles/
there are 4 small images, if we look at them, we see their urls like:
http://img.alpacine.com/carteles/0/444/44494-100.jpg
and when we get to the biger one (web-navigating) we can also see this:
http://img.alpacine.com/carteles/0/444/44494.jpg
So we just have to remove the "-100" part to get the bigger one.